Data.DRS.DataType

Description

DRS data type

Exported

data DRS

Discourse Representation Structure (DRS)

Constructors

LambdaDRS ((DRSVar, [DRSVar]), Int)
A lambda DRS (a variable, the set of referents to be applied to the DRS, and its argument position)

Merge DRS DRS
A merge between two DRSs

DRS [DRSRef] [DRSCon]
A DRS (a set of referents and a set of conditions)

type DRSVar = String

DRS variable

data DRSRef

DRS referent

Constructors

LambdaDRSRef ((DRSVar, [DRSVar]), Int)
A lambda DRS referent (a variable, the set of referents to be applied to the referent, and its argument position)

DRSRef DRSVar
A DRS referent

data DRSRel

DRS relation

Constructors

LambdaDRSRel ((DRSVar, [DRSVar]), Int)

DRSRel String

data DRSCon

DRS condition

Constructors

Rel DRSRel [DRSRef]
A relation defined on a set of referents

Neg DRS
A negated DRS

Imp DRS DRS
An implication between two DRSs

Or DRS DRS
A disjunction between two DRSs

Prop DRSRef DRS
A proposition DRS

Diamond DRS
A possible DRS

Box DRS
A necessary DRS