Data.DRS.Input.Boxer

Description

Converts Boxer's Prolog output to DRS

Exported

type PrologDRS = String

Data type for PrologDRS.

boxerToDRS :: String -> DRS

Converts Boxer's output into a DRS.

replaceLambdas :: PrologDRS -> Int -> PrologDRS

Replaces all lambda-variables in a PrologDRS by lambda-identifiers.

convertPrologVars :: PrologDRS -> [(String, String)] -> PrologDRS

Converts all Prolog Variables into new Variables.

Private

plDRSToDRS :: PrologDRS -> DRS

Converts a PrologDRS into a DRS.

parsePlRefs :: String -> [DRSRef]

Converts a String with Prolog referents into a set of DRSRefs.

parsePlCons :: String -> [DRSCon]

Converts a String with Prolog conditions into a set of DRSCons.

toDRSRef :: String -> DRSRef

Converts a String to a DRSRef, which may be a LambdaDRSRef.