Data.DRS.Input.String

Description

String to DRS

Exported

String to DRS conversion

stringToDRS :: String -> DRS

Transforms a String representation of a DRS into a DRS.

String representations for operators

opNegString :: [String]

Negation operators (case insensitive): !, not, neg.

opImpString :: [String]

Implication operators (case insensitive): imp, ->, =>, then.

opOrString :: [String]

Disjuction operators (case insensitive): v, or.

opBoxString :: [String]

Box operators (case insensitive): b, box, necessary.

opDiamondString :: [String]

Diamond operators (case insensitive): d, diamond, maybe.

Auxiliary functions for string parsing

data BracketType

Data type for brackets.

Constructors

Parentheses

Curly

Angle

Square

brackets :: BracketType -> (Char, Char)

Returns tuple with brackets from given BracketType.

felicitousBracketing :: String -> Bool

Checks if 'String s' contains the same number of opening and closing brackets of the same type.

dropOuterBrackets :: String -> String

Removes the outer brackets from a String.

takeUpToMatchingBracket :: BracketType -> String -> String

Returns part of String s between opening bracket ob of type bt and its matching closing bracket.

dropUpToMatchingBracket :: BracketType -> String -> String

Returns part of String s after opening bracket ob of type bt and its matching closing bracket.

replaceArrows :: String -> String

Replaces ASCII representations of arrows in String s by operator "imp".

splitOn :: Char -> String -> [String]

Modification of words, where String s is broken into parts based on delimiter c.

Private