Data.DRS.LambdaCalculus

Description

DRS lambda calculus; alpha conversion, beta reduction, and function composition

Exported

Type classes

class DRSAtom a

Type class for a DRSAtom, which is either a DRS or a DRSRef.

class AbstractDRS a

Type class for an AbstractDRS, which is either a resolved DRS, or an 'unresolved DRS' that takes a DRSAtom and yields an AbstractDRS.

Alpha Conversion

drsAlphaConvert :: DRS -> [(DRSRef, DRSRef)] -> DRS

Applies alpha conversion to a DRS on the basis of a conversion list for DRSRefs rs.

renameVar :: Eq a => a -> [(a, a)] -> a

Renames a variable v, iff v occurs in a variable conversion list. Otherwise, v is returned unmodified

Beta Reduction

drsBetaReduce :: (AbstractDRS a, DRSAtom b) => (b -> a) -> b -> a

Apply beta reduction on an AbstractDRS with a DRSAtom.

(<<@>>) :: (AbstractDRS a, DRSAtom b) => (b -> a) -> b -> a

Infix notation for drsBetaReduce

Function Composition

drsFunctionCompose :: (b -> c) -> (a -> b) -> a -> c

Apply function composition to two 'unresolved DRS's, yielding a new 'unresolved DRS'

(<<.>>) :: (b -> c) -> (a -> b) -> a -> c

Infix notation for drsFunctionCompose

DRS Purification

drsPurify :: DRS -> DRS

Converts a DRS into a pure DRS by purifying its DRSRefs, where:

A DRS is pure iff:

There are no occurrences of duplicate, unbound uses of the same DRSRef.

Private

DRS renaming

Purifying DRSRefs