Data.DRS.Properties

Description

DRS properties

Exported

isFOLDRS :: DRS -> Bool

Returns whether DRS d can be translated into a FOLForm.

isProperDRS :: DRS -> Bool

Returns whether DRS d is proper , where:

DRS d is proper iff

d does not contain any free variables

isPureDRS :: DRS -> Bool

Returns whether DRS d is pure, where:

DRS d is pure iff

d does not contain any otiose declarations of discourse referents (i.e., d does not contain any unbound, duplicate uses of referents).