Data.DRS.Variables

Description

DRS variables

Exported

Conversion

drsRefToDRSVar :: DRSRef -> DRSVar

Converts a DRSRef r into a DRSVar.

drsRelToString :: DRSRel -> String

Converts a DRSRef r into a DRSVar.

New Variables

newDRSRefs :: [DRSRef] -> [DRSRef] -> [DRSRef]

Returns a list of new DRSRefs, based on a list of old DRSRefs and a list of existing DRSRefs.

Variable Collections

drsUniverses :: DRS -> [DRSRef]

Returns the list of DRSRefs from all universes in a DRS.

drsVariables :: DRS -> [DRSRef]

Returns the list of all DRSRefs in a DRS (equals drsUniverses ++ drsFreeRefs).

drsLambdas :: DRS -> [(DRSVar, [DRSVar])]

Returns the ordered list of all lambda variables in a DRS.

Private

lambdas :: DRS -> [((DRSVar, [DRSVar]), Int)]

Returns the list of all lambda tuples in a DRS.

increase :: DRSVar -> DRSVar

Increases the index of a DRSVar by 1, or adds an index to it.