Data.PDRS.Input.Boxer

Description

Converts Boxer's Prolog output to PDRS

Exported

boxerToPDRS :: String -> PDRS

Converts Boxer's output into a PDRS.

Private

plPDRSToPDRS :: PrologDRS -> PDRS

Converts a PrologDRS into a PDRS.

plSDRSToPDRS :: String -> PDRS

Converts a PrologDRS of type sdrs into a PDRS.

XXX SDRSs are currently translated to PDRSs with same accessibility and spacial properties as SDRSs:

coordinated relation -> AMerge

subordinated relation -> Propositional conditions + MAP-constraint

parsePlRefs :: String -> [PRef]

Converts a String with Prolog referents into a set of PRefs.

parsePlCons :: String -> [PCon]

Converts a String with Prolog conditions into a set of PCons.

toPDRSRef :: String -> PDRSRef

Converts a String to a PDRSRef, which may be a LambdaPDRSRef.

bindPRefs :: PDRS -> [PRef] -> PDRS

Makes sure PDRS derived from Boxer does not contain free PRefs by adding MAPs where necessary. | XXX this function assumes that Boxer is right!