Data.PDRS.LambdaCalculus

Description

PDRS lambda calculus; alpha conversion, beta reduction, function composition, and `PDRS purification'

Exported

Alpha Conversion

pdrsAlphaConvert :: PDRS -> [(PVar, PVar)] -> [(PDRSRef, PDRSRef)] -> PDRS

Applies alpha conversion to a PDRS on the basis of conversion lists for PVars and PDRSRefs.

Beta Reduction

pdrsBetaReduce :: (AbstractPDRS a, PDRSAtom b) => (b -> a) -> b -> a

Apply beta reduction on an AbstractPDRS with a PDRSAtom.

(<<@>>) :: (AbstractPDRS a, PDRSAtom b) => (b -> a) -> b -> a

Infix notation for pdrsBetaReduce.

Function Composition

pdrsFunctionCompose :: (b -> c) -> (a -> b) -> a -> c

Apply function composition to two 'unresolved PDRS's, yielding a new 'unresolved PDRS'.

(<<.>>) :: (b -> c) -> (a -> b) -> a -> c

Infix notation for pdrsFunctionCompose.

PDRS Purification

pdrsPurify :: PDRS -> PDRS

Converts a PDRS into a pure PDRS by first purifying its projection variables, and then purifying its projected referents, where:

A PDRS is pure iff:

There are no occurrences of duplicate, unbound uses of the same variable (PVar or PDRSRef).

Private

Type classes

class PDRSAtom a

Type class for a PDRSAtom, which is either a PDRS or a PDRSRef.

class AbstractPDRS a

Type class for an AbstractPDRS, which is either a resolved PDRS, or an unresolved PDRS that takes a PDRSAtom and yields an AbstractPDRS.

PDRS renaming

renameSubPDRS :: PDRS -> PDRS -> [(PVar, PVar)] -> [(PDRSRef, PDRSRef)] -> PDRS

Applies alpha conversion to a PDRS sp, which is a sub-PDRS of the global PDRS gp, on the basis of conversion lists for PVars ps and PDRSRefs rs.

renameMAPs :: [MAP] -> PDRS -> PDRS -> [(PVar, PVar)] -> [MAP]

Applies alpha conversion to a list of MAPs m, on the basis of a conversion list for projection variables ps.

renameUniverse :: [PRef] -> PDRS -> PDRS -> [(PVar, PVar)] -> [(PDRSRef, PDRSRef)] -> [PRef]

Applies alpha conversion to a list of PRefs u, on the basis of a conversion list for PVars ps and PDRSRefs rs.

renamePCons :: [PCon] -> PDRS -> PDRS -> [(PVar, PVar)] -> [(PDRSRef, PDRSRef)] -> [PCon]

Applies alpha conversion to a list of PCons c in in local PDRS lp in global PDRS gp, on the basis of conversion lists for PVars ps and PDRSRefs rs.

renamePDRSRef :: PVar -> PDRSRef -> PDRS -> PDRS -> [(PDRSRef, PDRSRef)] -> PDRSRef

Applies alpha conversion to a list of projected referents u, in local PDRS lp in global PDRS gp,on the basis of a conversion list for projection variables ps and PDRS referents rs

renamePVar :: PVar -> PDRS -> PDRS -> [(PVar, PVar)] -> PVar

Converts a PVar into a new PVar in case it occurs bound in local PDRS lp in global PDRS gp.

Purifying PVars

purifyPVars :: (PDRS, [PVar]) -> PDRS -> (PDRS, [PVar])

Replaces duplicate uses of projection variables by new variables.

This function implements the following algorithm:

start with the global PDRS gp and add all free PVars in gp to the list of seen projection variables pvs (see pdrsPurify);

check the label l of the first atomic PDRS lp against pvs and, if necessary, alpha-convert lp replacing l for a new PVar;

add the label and all PVars from the universe and set of MAPs of pdrs to the list of seen projection variables pvs;

go through all conditions of pdrs, while continually updating pvs.

Purifying PRefs

purifyPRefs :: PDRS -> PDRS -> [(PRef, PRef)] -> PDRS

Replaces duplicate uses of projection referents by new PRefs, where:

Given conversion pair (pr',npr), pr is replaced by npr iff

pr equals pr', or

pr and pr' have the same referent and pr is bound by pr' in global PDRS gp, or

pr and pr' have the same referent and both occur free in subordinated contexts in gp.

unboundDupPRefs :: PDRS -> PDRS -> [PRef] -> ([PRef], [PRef])

Returns a tuple of existing PRefs eps and unbound duplicate PRefs dps in a PDRS, based on a list of seen PRefs prs, where:

pr = (PRef p r) is duplicate in PDRS gp iff

There exists a p', such that pr' = (PRef p' r) is an element of prs, and pr and pr' are independent.

independentPRefs :: PRef -> [PRef] -> PDRS -> PDRS -> Bool

Returns whether a PRef pr is independent based on a list of PRefs prs, where:

pr = (PRef p r) is not independent w.r.t. prs iff

pr is bound by any PRef in prs; and

pr occurs free and there is some element of prs that is accessible from pr. (NB. this only holds if both pr and pr' occur free in accessible contexts, in which case they are not independent).

pr occurs free and there is some element of prs that may become accessible from pr at some later point, because its projection site is undetermined with respect to the projection site of pr. This is the case if the projection site occurs free (i.e., has no path to the global context).