Data.PDRS.Show

Description

PDRS pretty printing

Exported

Show PDRS

class ShowablePDRS p where

Typeclass for showablePDRSs, that are unresolved.

Minimal complete definition

resolve

Methods

resolve :: p -> Int -> Int -> PDRS

data PDRSNotation p

PDRS notations.

Constructors

Set p
Set notation

Linear p
Linear notation

Boxes p
Box notation

Debug p
Debug notation

showPDRS :: PDRSNotation PDRS -> String

Shows a PDRS.

printPDRS :: PDRS -> IO ()

Prints a PDRS.

Show Merges

showAMerge :: PDRS -> PDRS -> String

Shows an assertive merge between PDRS p1 and PDRS p.

printAMerge :: PDRS -> PDRS -> IO ()

Prints an assertive merge between PDRS p1 and PDRS p.

showPMerge :: PDRS -> PDRS -> String

Shows a projective merge between PDRS p1 and PDRS p.

printPMerge :: PDRS -> PDRS -> IO ()

Print a projective merge between PDRS p1 and PDRS p.

Show Beta Reduction

showPDRSBetaReduct :: ShowablePDRS p => (PDRS -> p) -> PDRS -> String

Shows the beta reduction of an 'unresolved PDRS' p1 with a PDRS p2.

printPDRSBetaReduct :: ShowablePDRS p => (PDRS -> p) -> PDRS -> IO ()

Prints the beta reduction of an 'unresolved PDRS' p1 with a PDRS p2.

showPDRSRefBetaReduct :: ShowablePDRS p => (PDRSRef -> p) -> PDRSRef -> String

Shows the beta reduction of an 'unresolved PDRS' p with a PDRS referent r.

printPDRSRefBetaReduct :: ShowablePDRS p => (PDRSRef -> p) -> PDRSRef -> IO ()

Prints the beta reduction of an 'unresolved PDRS' p with a PDRSRef r.

Private

Operators

opAMerge :: String

Assertive merge symbol.

opPMerge :: String

Projective merge symbol.

modPointer :: String

Pointer symbol.

modEquals :: String

Equals symbol.

modWeakSubord :: String

Weak subordination symbol.

modStrictSubord :: String

Weak subordination symbol.

Notations for showing PDRSs

showPDRSBox :: PDRS -> String

Show a PDRS in Box notation.

showPDRSLinear :: PDRS -> String

Show a PDRS in Linear notation.

showPDRSSet :: PDRS -> String

Show a PDRS in Set notation.

showPDRSDebug :: PDRS -> String

Show a PDRS in Debug notation.

Showing the subparts of a PDRS

showHeaderLine :: Int -> PVar -> String

Shows a horizontal line of width l that label pl in its center.

showUniverse :: [PRef] -> String

Shows the universe u of a PDRS.

showUniverseTuples :: [PRef] -> String

Shows the universe u of a PDRS as tuples.

showConditions :: [PCon] -> String

Shows the projected conditions c of a PDRS.

showMAPs :: [MAP] -> String

Shows the MAPs m of a PDRS.

showMAPsTuples :: [MAP] -> String

Shows the MAPs m of a PDRS as tuples.

showPDRSLambdas :: PDRS -> String

Shows lambda abstractions over PDRS p.

Orphan instances

Show PDRS
Derive and instance of the Show typeclass for PDRS.

Methods

showsPrec :: Int -> PDRS -> ShowS

show :: PDRS -> String

showList :: [PDRS] -> ShowS

ShowablePDRS p => Show ((PDRSRef -> PDRS) -> p)

Methods

showsPrec :: Int -> ((PDRSRef -> PDRS) -> p) -> ShowS

show :: ((PDRSRef -> PDRS) -> p) -> String

showList :: [(PDRSRef -> PDRS) -> p] -> ShowS

ShowablePDRS p => Show (PDRSRel -> p)

Methods

showsPrec :: Int -> (PDRSRel -> p) -> ShowS

show :: (PDRSRel -> p) -> String

showList :: [PDRSRel -> p] -> ShowS

ShowablePDRS p => Show (PDRSRef -> p)
Derive appropriate instances of Show for ShowablePDRSs.

Methods

showsPrec :: Int -> (PDRSRef -> p) -> ShowS

show :: (PDRSRef -> p) -> String

showList :: [PDRSRef -> p] -> ShowS

ShowablePDRS p => Show (PDRS -> p)

Methods

showsPrec :: Int -> (PDRS -> p) -> ShowS

show :: (PDRS -> p) -> String

showList :: [PDRS -> p] -> ShowS