Data.PDRS.Structure

Description

Structural operations on PDRSs

Exported

pdrsLabel :: PDRS -> PVar

Returns the label of a PDRS.

pdrsLabels :: PDRS -> [PVar]

Returns all the labels in a PDRS.

pdrsUniverse :: PDRS -> [PRef]

Returns the universe of a PDRS

pdrsUniverses :: PDRS -> [PRef]

Returns the list of projected referents in all universes of a PDRS.

pdrsMAPs :: PDRS -> [MAP]

Returns the list of MAPs of a PDRS.

emptyPDRS :: PDRS -> PDRS

Returns an empty PDRS, if possible with the same label as PDRS p.

isLambdaPDRS :: PDRS -> Bool

Returns whether a PDRS is entirely a LambdaPDRS (at its top-level)

isMergePDRS :: PDRS -> Bool

Returns whether a PDRS is an AMerge or PMerge (at its top-level)

isResolvedPDRS :: PDRS -> Bool

Returns whether a PDRS is resolved (containing no unresolved merges or lambdas)

isSubPDRS :: PDRS -> PDRS -> Bool

Returns whether PDRS p1 is a direct or indirect sub-PDRS of PDRS p2.