Data.PDRS.Translate

Description

PDRS accommodation, PDRS to FOL translation, and PDRS to DRS translation

Exported

pdrsToFOL :: PDRS -> FOLForm

Translates a PDRS into a FOLForm.

pdrsToDRS :: PDRS -> DRS

Translates a PDRS into a DRS.

Private

movePContent :: PDRS -> PDRS -> PDRS -> PDRS

Moves projected content in PDRS to its interpretation site in PDRS lp based on global PDRS gp.

insertPRefs :: [PRef] -> PDRS -> PDRS -> PDRS

Inserts projected referents prs in a PDRS p at the correct interpretation site, based on the global PDRS gp.

insertPCon :: PCon -> PDRS -> PDRS -> PDRS

Inserts projected condition pc in PDRS p at its interpretation site, based on the global PDRS gp.

stripPVars :: PDRS -> DRS

Strips projection variables from a PDRS p, resulting in a DRS.