Data.DRS.LambdaCalculus

Description

DRS lambda calculus; alpha conversion, beta reduction, and function composition

Exported

Alpha Conversion

drsAlphaConvert :: DRS -> [(DRSRef, DRSRef)] -> DRS

Applies alpha conversion to a DRS on the basis of a conversion list for DRSRefs rs.

renameVar :: Eq a => a -> [(a, a)] -> a

Renames a variable v, iff v occurs in a variable conversion list. Otherwise, v is returned unmodified

Beta Reduction

drsBetaReduce :: (AbstractDRS a, DRSAtom b) => (b -> a) -> b -> a

Apply beta reduction on an AbstractDRS with a DRSAtom.

(<<@>>) :: (AbstractDRS a, DRSAtom b) => (b -> a) -> b -> a

Infix notation for drsBetaReduce

Function Composition

drsFunctionCompose :: (b -> c) -> (a -> b) -> a -> c

Apply function composition to two 'unresolved DRS's, yielding a new 'unresolved DRS'

(<<.>>) :: (b -> c) -> (a -> b) -> a -> c

Infix notation for drsFunctionCompose

DRS Purification

drsPurify :: DRS -> DRS

Converts a DRS into a pure DRS by purifying its DRSRefs, where:

A DRS is pure iff:

There are no occurrences of duplicate, unbound uses of the same DRSRef.

Private

Type classes

class DRSAtom a

Type class for a DRSAtom, which is either a DRS or a DRSRef.

class AbstractDRS a

Type class for an AbstractDRS, which is either a resolved DRS, or an 'unresolved DRS' that takes a DRSAtom and yields an AbstractDRS.

DRS renaming

renameSubDRS :: DRS -> DRS -> [(DRSRef, DRSRef)] -> DRS

Applies alpha conversion to a DRS sd, which is a sub-DRS of the global DRS gd, on the basis of a conversion list for DRSRefs rs.

renameRefs :: [DRSRef] -> [(DRSRef, DRSRef)] -> [DRSRef]

Applies alpha conversion to a list of DRSRefs u, on the basis of a conversion list rs

renameCons :: DRS -> DRS -> [(DRSRef, DRSRef)] -> [DRSCon]

Applies alpha conversion to the conditions in a DRS sd, which is a sub-DRS of gd, on the basis of a conversion list for DRSRefs rs.

Purifying DRSRefs

purifyRefs :: (DRS, [DRSRef]) -> DRS -> (DRS, [DRSRef])

Replaces duplicate uses of DRSRefs by new DRSRefs.

This function implements the following algorithm:

start with the global DRS gd and add all free PVars in gd to the list of seen referents rs (see drsPurify);

check the universe u of the first atomic DRS ld against rs and, if necessary, alpha-convert ld replacing duplicates for new DRSRefs in u;

add the universe of ld to the list of seen DRSRefs rs;

go through all conditions of ld, while continually updating rs.