Data.DRS.Show

Description

DRS pretty printing

Exported

Show DRS

class ShowableDRS d where

Typeclass for showableDRSs, that are unresolved.

Methods

drsResolve :: d -> Int -> Int -> DRS

data DRSNotation d

DRS notations.

Constructors

Set d
Set notation

Linear d
Linear notation

Boxes d
Box notation

Debug d
Debug notation

showDRS :: DRSNotation DRS -> String

Shows a DRS.

printDRS :: DRS -> IO ()

Prints a DRS.

Show Merge

showMerge :: DRS -> DRS -> String

Shows a merge between two DRSs.

printMerge :: DRS -> DRS -> IO ()

Prints a merge between two DRSs.

Show Beta Reduction

showDRSBetaReduct :: ShowableDRS d => (DRS -> d) -> DRS -> String

Shows the beta reduction of an 'unresolved DRS' d1 with a DRS d2.

printDRSBetaReduct :: ShowableDRS d => (DRS -> d) -> DRS -> IO ()

Prints the beta reduction of an 'unresolved DRS' d1 with a DRS d2.

showDRSRefBetaReduct :: ShowableDRS d => (DRSRef -> d) -> DRSRef -> String

Shows the beta reduction of an 'unresolved DRS' d with a DRSRef r.

printDRSRefBetaReduct :: ShowableDRS d => (DRSRef -> d) -> DRSRef -> IO ()

Prints the beta reduction of an 'unresolved DRS' d with a DRSRef r.

Operators

opNeg :: String

Negation symbol

opImp :: String

Implication symbol

opOr :: String

Disjunction symbol

opDiamond :: String

Diamond symbol

opBox :: String

Box symbol

opLambda :: String

Lambda symbol

Box Construction

boxTopLeft :: Char

Top left corner symbol

boxTopRight :: Char

Top right corner symbol

boxBottomLeft :: Char

Bottom left corner symbol

boxBottomRight :: Char

Bottom right corner symbol

boxMiddleLeft :: Char

Middle left corner symbol

boxMiddleRight :: Char

Middle right corner symbol

boxHorLine :: Char

Horizontal line symbol

boxVerLine :: Char

Vertical line symbol

showConcat :: String -> String -> String

Shows the line by line concatenation of two Strings.

showContent :: Int -> String -> String

Shows the content of a DRS box surrounded by vertical bars.

showHorizontalLine :: Int -> Char -> Char -> String

Shows a horizontal line of length l with left corner symbol lc and right corner symbol rc.

showModifier :: String -> Int -> String -> String

Shows a modifier m at line number p in front of String s.

showPadding :: String -> String

Adds two lines of whitespace to the beginning of String s.

Private

Notations for showing DRSs

Showing the subparts of a DRS

Orphan instances

Show DRS
Derive an instance of the Show typeclass for DRS.

Instance details

Methods

showsPrec :: Int -> DRS -> ShowS

show :: DRS -> String

showList :: [DRS] -> ShowS

ShowableDRS d => Show (DRS -> d)

Instance details

Methods

showsPrec :: Int -> (DRS -> d) -> ShowS

show :: (DRS -> d) -> String

showList :: [DRS -> d] -> ShowS

ShowableDRS d => Show (DRSRef -> d)
Derive appropriate instances of Show for ShowableDRSs.

Instance details

Methods

showsPrec :: Int -> (DRSRef -> d) -> ShowS

show :: (DRSRef -> d) -> String

showList :: [DRSRef -> d] -> ShowS

ShowableDRS d => Show (DRSRel -> d)

Instance details

Methods

showsPrec :: Int -> (DRSRel -> d) -> ShowS

show :: (DRSRel -> d) -> String

showList :: [DRSRel -> d] -> ShowS

ShowableDRS d => Show ((DRSRef -> DRS) -> d)

Instance details

Methods

showsPrec :: Int -> ((DRSRef -> DRS) -> d) -> ShowS

show :: ((DRSRef -> DRS) -> d) -> String

showList :: [(DRSRef -> DRS) -> d] -> ShowS