Data.PDRS.DataType

Description

PDRS data type

Exported

data PDRS

Projective Discourse Representation Structure.

Constructors

LambdaPDRS ((DRSVar, [DRSVar]), Int)
A lambda PDRS (a variable, the set of referents to be applied to the PDRS, and its argument position)

AMerge PDRS PDRS
An assertive merge between two PDRSs

PMerge PDRS PDRS
A projective merge between two PDRSs

PDRS PVar [MAP] [PRef] [PCon]
A PDRS, consisting of a PVar (a label), a set of MAPs, a set of PRefs, and a set of PCons

type PVar = Int

Projection variable (a label or pointer).

type MAP = (PVar, PVar)

Minimally Accessible PDRS, represented as a tuple between a PVar and another PVar that is minimally accessible from the first PVar.

data PRef

A projected referent, consisting of a PVar and a PDRSRef.

Constructors

PRef PVar PDRSRef

data PDRSRef

A PDRS referent.

Constructors

LambdaPDRSRef ((DRSVar, [DRSVar]), Int)
A lambda PDRS referent (a variable, the set of referents to be applied to the referent, and its argument position)

PDRSRef DRSVar
A PDRS referent

data PDRSRel

PDRS relation

Constructors

LambdaPDRSRel ((DRSVar, [DRSVar]), Int)

PDRSRel String

data PCon

A projected condition, consisting of a PVar and a PDRSCon.

Constructors

PCon PVar PDRSCon

data PDRSCon

A PDRS condition.

Constructors

Rel PDRSRel [PDRSRef]
A relation defined on a set of referents

Neg PDRS
A negated PDRS

Imp PDRS PDRS
An implication between two PDRSs

Or PDRS PDRS
A disjunction between two PDRSs

Prop PDRSRef PDRS
A proposition PDRS

Diamond PDRS
A possible PDRS

Box PDRS
A necessary PDRS