Data.PDRS.Variables

Description

PDRS variables

Exported

Conversion

pdrsRefToDRSRef :: PDRSRef -> DRSRef

Converts a PDRSRef into a DRSRef

pdrsRefToDRSVar :: PDRSRef -> DRSVar

Directly converts a PDRSRef into a DRSVar

drsRefToPDRSRef :: DRSRef -> PDRSRef

Converts a DRSRef to a PDRSRef

pdrsRelToDRSRel :: PDRSRel -> DRSRel

Converts a PDRSRel into a DRSRel

pdrsRelToString :: PDRSRel -> String

Directly converts a PDRSRel into a String

New Variables

newPVars :: [PVar] -> [PVar] -> [PVar]

Returns a list of new projection variables for a list of old PVars opvs, based on a list of existing PVars epvs.

newPDRSRefs :: [PDRSRef] -> [PDRSRef] -> [PDRSRef]

Returns a list of new referents for a list of old PDRSRefs ors, based on a list of existing PDRSRefs ers.

newPRefs :: [PRef] -> [PDRSRef] -> [PRef]

Returns a list of new projected referents for a list of old PRefs, based on a list of existing PVars prs and existing PDRSRefs ers.

Variable Collections

pdrsVariables :: PDRS -> [PDRSRef]

Returns the list of all variables in a PDRS

pdrsPVars :: PDRS -> [PVar]

Returns the list of all PVars in a PDRS

pdrsLambdas :: PDRS -> [((DRSVar, [DRSVar]), Int)]

Returns the list of all lambda tuples in a PDRS

pdrsLambdaVars :: PDRS -> [(DRSVar, [DRSVar])]

Returns the list of all lambda variables in a PDRS